+++ lib/zaikio/mission_control/jobs/folder.rb
@@ -1,0 +1,9 @@
+module Zaikio
+  module MissionControl
+    module Jobs
+      class Folder < Base
+        has_many_parts :folders, required: true
+      end
+    end
+  end
+end