+++ lib/zaikio/mission_control/jobs/folding_card.rb
@@ -1,0 +1,9 @@
+module Zaikio
+  module MissionControl
+    module Jobs
+      class FoldingCard < Base
+        has_many_parts :folding_cards, required: true
+      end
+    end
+  end
+end